+++ src/components/Message/renderText/index.ts
@@ -1,4 +1,6 @@
 export { MentionProps } from './Mention';
-export { emojiMarkdownPlugin, mentionsMarkdownPlugin } from './rehypePlugins';
 export { escapeRegExp, matchMarkdownLinks, messageCodeBlocks } from './regex';
+export * from './rehypePlugins';
+export * from './remarkPlugins';
 export * from './renderText';
+export { HNode } from './types';