+++ src/components/Message/renderText/remarkPlugins/index.ts
@@ -1,0 +1,2 @@
+export * from './htmlToTextPlugin';
+export * from './keepLineBreaksPlugin';