+++ src/components/Message/renderText/types.ts
@@ -1,0 +1,3 @@
+import { Content, Root } from 'hast';
+
+export type HNode = Content | Root;